+++ src/kernel/mpipeline.ml
@@ -154,18 +154,26 @@ module Reader_phase = struct
   module Fingerprint = struct
     type t =
       { source_digest : Msource.Digest.t;
-        for_completion : Msource.position option
+        for_completion : Msource.position option;
+        config : Mconfig.t
       }
 
-    let make { source = source, _; for_completion; _ } =
-      Ok { source_digest = Msource.Digest.make source; for_completion }
+    let make { source = source, _; for_completion; config; _ } =
+      Ok { source_digest = Msource.Digest.make source; for_completion; config }
 
     let equal
-        { source_digest = source_digest_1; for_completion = for_completion_1 }
-        { source_digest = source_digest_2; for_completion = for_completion_2 } =
+        { source_digest = source_digest_1;
+          for_completion = for_completion_1;
+          config = config_1
+        }
+        { source_digest = source_digest_2;
+          for_completion = for_completion_2;
+          config = config_2
+        } =
       Msource.Digest.equal source_digest_1 source_digest_2
       && Stdlib.Option.equal Msource.equal_position for_completion_1
            for_completion_2
+      && Cache.key config_1 = Cache.key config_2
   end
 end